+++ lib/attributor/types/date_time.rb
@@ -31,7 +31,7 @@ module Attributor
     end
 
     def self.dump(value,**opts)
-      value.iso8601
+      value && value.iso8601
     end