+++ lib/attributor/types/time.rb
@@ -34,7 +34,7 @@ module Attributor
     end
 
     def self.dump(value,**opts)
-      value.iso8601
+      value && value.iso8601
     end