+++ spec/types/date_time_spec.rb
@@ -18,6 +18,11 @@ describe Attributor::DateTime do
     it 'is formatted correctly' do
       value.should match(/\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[\+-]\d{2}:\d{2}/)
     end
+    context 'nil values' do
+      it 'should be nil' do
+        type.dump(nil).should be_nil 
+      end
+    end
   end